+++ src/controllers/loginControllers.js
@@ -1,9 +1,30 @@
 const { User } = require('../db');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
-const mailer = require('../functions/mailer');
+// const mailer = require('../functions/mailer');
 const secretKey = 'mi_secreto';
+const nodemailer = require("nodemailer");
 // const secretKey = process.env.SECRET_KEY;
+
+async function mailer(email) {
+    const transporter = nodemailer.createTransport({
+        host: "smtp.gmail.com",
+        port: 465,
+        secure: true,
+        auth: {
+            user: "instachef2@gmail.com",
+            pass: "ylhfxhukvdnoleam",
+        },
+    });
+
+    let info = await transporter.sendMail({
+        from: '"👩‍🍳InstaChef👨‍🍳" <instachef2@gmail.com>',
+        to: email,
+        subject: "Registro exitoso",
+        text: "Bienvenido a InstaChef",
+        html: "<b>Bienvenido a InstaChef. Gracias por elegirnos y ¡que disfrute de su comida!</b>",
+    });
+}
 
 const signUp = async (req, res) => {
   try {
